kernel: Boost whenever a zygote-forked process becomes a top app.
Boost to the max for 1250 ms whenever the top app changes, which
improves app launch speeds and addresses jitter when switching between
apps. A check to make sure that the top-app's parent is zygote ensures
that a user-facing app is indeed what's added to the top app task group,
since app processes are forked from zygote.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: Ia3650af54d356de335832c4c5e2576705fe09da1
Signed-off-by: SparXFusion <s2234nadar@gmail.com>
Signed-off-by: techyminati <sinha.aryan03@gmail.com>