cgroup: Implement Optionally killable freezer
Adds a configurable option to allow killing frozen tasks in a way
similar to cgroups v2. The status of frozen tasks is changed to
TASK_INTERRUPTIBLE and their PF_FROZEN flag is reset on pending fatal
signals.

This behavior is enabled via a run-time configurable option, so that
the pre-existing behavior is preserved by default.

Change-Id: Ic86a2c2652ab4b374cd636faf48bc55065858b3f
Bug: b/143308662
Test: manual
Signed-off-by: Marco Ballesio <balejs@google.com>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: SparXFusion <s2234nadar@gmail.com>
Signed-off-by: techyminati <sinha.aryan03@gmail.com>